LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-stretch-properties-dynamic-001-expected.txt


FAIL minsize assert_approx_equals: attach expected 100 +/- 1 but got 25.015625
FAIL maxsize assert_approx_equals: attach expected 100 +/- 1 but got 25.015625
FAIL largeop assert_approx_equals: set true expected 50 +/- 1 but got 26
FAIL symmetric assert_approx_equals: set true expected 150 +/- 1 but got 25.015625
PASS stretchy
⥯
⥯
⥯

⥯
⥯
⥯

⥯
⥯

⥯
⥯

⥯
⥯
